move predicate expansion to predicate package
This moves `(*searchResolver).expandPredicates()` out of graphqlbackend
and to `internal/search/predicate:Expand()`. Just continuing our
slimming down of `searchResolver`.
