+++ internal/search/predicate/substitute.go
@@ -1,14 +1,83 @@
 package predicate
 
 import (
-	"github.com/grafana/regexp"
+	"context"
+	"sync"
 
+	"github.com/grafana/regexp"
+	"golang.org/x/sync/errgroup"
+
+	"github.com/sourcegraph/sourcegraph/internal/database"
+	"github.com/sourcegraph/sourcegraph/internal/search/job"
 	"github.com/sourcegraph/sourcegraph/internal/search/query"
 	"github.com/sourcegraph/sourcegraph/internal/search/result"
+	"github.com/sourcegraph/sourcegraph/internal/search/streaming"
+	"github.com/sourcegraph/sourcegraph/internal/trace"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 var ErrNoResults = errors.New("no results returned for predicate")
+
+// Expand takes a query plan, and replaces any predicates with their expansion. The returned plan
+// is guaranteed to be predicate-free.
+func Expand(ctx context.Context, db database.DB, jobArgs *job.Args, oldPlan query.Plan) (_ query.Plan, err error) {
+	tr, ctx := trace.New(ctx, "ExpandPredicates", "")
+	defer func() {
+		tr.SetError(err)
+		tr.Finish()
+	}()
+
+	var (
+		mu      sync.Mutex
+		newPlan = make(query.Plan, 0, len(oldPlan))
+	)
+	g, ctx := errgroup.WithContext(ctx)
+
+	for _, q := range oldPlan {
+		q := q
+		g.Go(func() error {
+			predicatePlan, err := Substitute(q, func(plan query.Plan) (result.Matches, error) {
+				children := make([]job.Job, 0, len(plan))
+				for _, basicQuery := range plan {
+					child, err := job.ToEvaluateJob(jobArgs, basicQuery)
+					if err != nil {
+						return nil, err
+					}
+					children = append(children, child)
+				}
+
+				agg := streaming.NewAggregatingStream()
+				_, err = job.NewOrJob(children...).Run(ctx, db, agg)
+				if err != nil {
+					return nil, err
+				}
+				return agg.Results, nil
+			})
+			if errors.Is(err, ErrNoResults) {
+				// The predicate has no results, so neither will this basic query
+				return nil
+			}
+			if err != nil {
+				// Fail if predicate processing fails.
+				return err
+			}
+
+			mu.Lock()
+			defer mu.Unlock()
+
+			if predicatePlan != nil {
+				// If the predicate generated a new plan, use that
+				newPlan = append(newPlan, predicatePlan...)
+			} else {
+				// Otherwise, just use the original basic query
+				newPlan = append(newPlan, q)
+			}
+			return nil
+		})
+	}
+
+	return newPlan, g.Wait()
+}
 
 // Substitute replaces predicates that generate plans and substitutes them to create a new Plan.
 func Substitute(q query.Basic, evaluate func(query.Plan) (result.Matches, error)) (query.Plan, error) {
